Many many updates
Tbh I lost track at this point, but the main updates are TP autosolvers, bundled manuals, and changing the functionality of the audio and skip features to function more like the other identification mods out there. For example VTuber only shows pictures so I changed it to function like Plant and Pickup ID where it cannot be skipped in anyway. For Piece though you can still skip but it will only allow for this when its not a strike or solve animation. Other small updates consist of lowering the text on the buttons so they don't float, making the text in the input box not viewable from the other side of the bomb, and many code optimizations.